+++ helper/filesystem/file.go
@@ -18,9 +18,14 @@ type File struct {
 // The file is not readable anymore once saved as its FileReader has already been
 // closed.
 //
+// Creates directories if needed.
+//
 // Returns the actual file name.
 func (file *File) Save(path string, name string) string {
 	name = timestampFileName(name)
+	if err := os.MkdirAll(path, os.ModePerm); err != nil {
+		panic(err)
+	}
 	writer, err := os.OpenFile(path+string(os.PathSeparator)+name, os.O_WRONLY|os.O_CREATE, 0660)
 	if err != nil {
 		panic(err)