+++ docs_src/src/guide/changelog.md
@@ -17,6 +17,7 @@ meta:
 - Added a way to customize the request's body field used by `JWTController` for the authentication process. (By default, "username" and "password" are used)
 - Added a new validation rule: `unique`.
 - Added `helper.Only()`.
+- `filesystem.File.Save()` now creates directories if needed.
 
 ## v3.1.0